Tests: Add unit test structure for WP-CLI class
We need to manually load WP-CLI when we bootstrap PHPUnit so the relevant classes are loaded.

Strictly speaking when it comes to unit tests, we shouldn’t need to bootstrap any dependency like this, as we aren’t testing WP-CLI and so don’t care about it, but as our class extends from a WP-CLI class, there isn’t a way to swap that out for a mocked class without some over-engineered class SPL aliasing in our actual plugin.

See https://polevaultweb.com/2016/11/unit-testing-custom-wp-cli-commands/

Includes a simple unit test to confirm that the custom WP-CLI class can be instantiated.