Insert redirect: allow return of inserted post ID
The `insert_legacy_redirect()` method will return "True if inserted; false if not permitted; otherwise error upon validation issue.".

Rather than just returning boolean true when a redirect has been successfully inserted, we can instead return the ID of the entry so that further work may be done with it by other code.

If we updated the `true` return, then this may break existing `===` checks. Instead, we add a new param to the method and which can dictate whether true or the post ID is returned on success.

Fixes #99.