Unit Tests: Disable code coverage by default
Including code coverage when running unit tests increases the time from 500ms to around 4.5s, so disable it by default, but add a Composer script to allow it to be easily run locally as needed.

The `echo coverage/index.html` allows a CMD-click in some terminals to immediately open the HTML report.